+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Set_Parameters"></activity>
+        <activity android:name=".Sign_up_player"></activity>
+        <activity android:name=".Sign_up_admin" />
+        <activity android:name=".Set_Parameters" />
         <activity
             android:name=".Player_Sign_in"
             android:label="@string/title_activity_player__sign_in"
