+++ app/src/main/java/com/example/myapplication/MainActivity.java
@@ -27,7 +27,7 @@ public class MainActivity extends AppCompatActivity {
            public void onClick(View v) {
 
                //put admin auth stuff
-               Intent intent = new Intent(context, Login_Admin.class);
+               Intent intent = new Intent(context, Sign_up_admin.class);
                context.startActivity(intent);
                }});
 
@@ -36,7 +36,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 //put stuff for player auth
-                Intent intent = new Intent(context, Player_Sign_in.class);
+                Intent intent = new Intent(context, Sign_up_player.class);
                 context.startActivity(intent);
 
             }
